agri-strat/requirements.txt
torch
matplotlib
netCDF4
numpy
pandas
pycocotools
lightning
scikit-learn
scikit-multilearn
seaborn
tensorboard
torchmetrics
torchvision
tqdm
xarray
wandb
dataclasses-json
rich
typing_extensions
packaging
pyyaml
pyarrow
